requirements.txt
Flask
tensorflow==2.15.0
numpy
pillow
gunicorn
